Left panel: split up large code files (#9967)
