+++ GitUI/BranchTreePanel/RepoObjectsTree.DummyNode.cs
@@ -1,0 +1,13 @@
+﻿namespace GitUI.BranchTreePanel
+{
+    partial class RepoObjectsTree
+    {
+        // Used temporarily to facilitate building a tree
+        private class DummyNode : Node
+        {
+            public DummyNode() : base(null)
+            {
+            }
+        }
+    }
+}
